requirements.txt
# LinguaFlow Learning Lab — Streamlit app + all project tabs (e.g. Streamlit Community Cloud).
# Install from repo root: pip install -r requirements.txt

streamlit>=1.40.0
python-dotenv>=1.0.0

langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-anthropic>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
langgraph>=0.4.0
langgraph-checkpoint-sqlite>=2.0.0
langsmith>=0.2.0
deepagents>=0.4.0

chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
